refactor(daemon): group design lifecycle modules into design/
