+++ apps/daemon/src/design/claude-design-import.ts
@@ -1,7 +1,7 @@
 import { mkdir, readFile, writeFile } from 'node:fs/promises';
 import path from 'node:path';
 import { inflateRawSync } from 'node:zlib';
-import { validateProjectPath } from './projects.js';
+import { validateProjectPath } from '../projects.js';
 
 const EOCD_SIG = 0x06054b50;
 const CENTRAL_SIG = 0x02014b50;
